backend/requirements.txt
fastapi
uvicorn
pandas
numpy
pydantic
python-multipart
openpyxl
Pillow